Validators added

+++ Igrm.OpenSkyApi/Models/Request/AllStateVectorsRequestModel.cs
@@ -38,7 +38,7 @@ namespace Igrm.OpenSkyApi.Models.Request
         /// <summary>
         /// The time in seconds since epoch (Unix time stamp to retrieve states for. Current time will be used if omitted.
         /// </summary>
-        public int? Time { get; set; }
+        public long? Time { get; set; }
 
         /// <summary>
         /// One or more ICAO24 transponder addresses represented by a hex string (e.g. abc9f3). To filter multiple ICAO24 append the property once for each address. If omitted, the state vectors of all aircraft are returned.
