+++ Igrm.OpenSkyApi/Models/Response/DeparturesByAirportResponseModel.cs
@@ -13,7 +13,7 @@ namespace Igrm.OpenSkyApi.Models.Response
         ///<summary>
         ///Estimated time of departure for the flight as Unix time (seconds since epoch).
         ///</summary>
-        public int Firstseen { get; set; }
+        public long Firstseen { get; set; }
         ///<summary>
         ///ICAO code of the estimated departure airport. Can be null if the airport could not be identified.
         ///</summary>
@@ -21,7 +21,7 @@ namespace Igrm.OpenSkyApi.Models.Response
         ///<summary>
         ///Estimated time of arrival for the flight as Unix time (seconds since epoch)
         ///</summary>
-        public int Lastseen { get; set; }
+        public long Lastseen { get; set; }
         ///<summary>
         ///ICAO code of the estimated arrival airport. Can be null if the airport could not be identified.
         ///</summary>
@@ -33,27 +33,27 @@ namespace Igrm.OpenSkyApi.Models.Response
         ///<summary>
         ///Horizontal distance of the last received airborne position to the estimated departure airport in meters
         ///</summary>
-        public int Estdepartureairporthorizdistance { get; set; }
+        public long Estdepartureairporthorizdistance { get; set; }
         ///<summary>
         ///Vertical distance of the last received airborne position to the estimated departure airport in meters
         ///</summary>
-        public int Estdepartureairportvertdistance { get; set; }
+        public long Estdepartureairportvertdistance { get; set; }
         ///<summary>
         ///Horizontal distance of the last received airborne position to the estimated arrival airport in meters
         ///</summary>
-        public int Estarrivalairporthorizdistance { get; set; }
+        public long Estarrivalairporthorizdistance { get; set; }
         ///<summary>
         ///Vertical distance of the last received airborne position to the estimated arrival airport in meters
         ///</summary>
-        public int Estarrivalairportvertdistance { get; set; }
+        public long Estarrivalairportvertdistance { get; set; }
         ///<summary>
         ///Number of other possible departure airports. These are airports in short distance to estDepartureAirport.
         ///</summary>
-        public int Departureairportcandidatescount { get; set; }
+        public long Departureairportcandidatescount { get; set; }
         ///<summary>
         ///Number of other possible departure airports. These are airports in short distance to estArrivalAirport.
         ///</summary>
-        public int Arrivalairportcandidatescount { get; set; }
+        public long Arrivalairportcandidatescount { get; set; }
 
     }
 }
